pyproject.toml
[tool.poetry]
name = "cleaned-multispeaker-transcripts"
version = "0.1.0"
description = ""
authors = ["Avinash Jain <avinashkalpitjain@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
assemblyai = "^0.22.0"
openai = "0.27.8"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
